Fixes bugs

+++ src/index.js
@@ -2,6 +2,7 @@ const path = require('path');
 const https = require('follow-redirects').https;
 const fs = require('fs-extra');
 const extract = require('extract-zip');
+const rp = require('request-promise');
 
 const { WIDEVINECDM_VERSION } = require('./constants');
 
@@ -20,18 +21,40 @@ const downloadAsync = (app, dest, platform = process.platform, arch = process.ar
   const fileName = `widevinecdm_${platform}_${arch}.zip`;
   const localZipPath = `${app.getPath('temp')}/widevinecdm-${process.pid}-${Date.now()}.zip`;
 
-  return new Promise((resolve, reject) => {
-    const file = fs.createWriteStream(localZipPath);
-    https.get(`https://github.com/webcatalog/electron-widevinecdm/releases/download/latest/${fileName}`, (response) => {
-      response.pipe(file);
-      file.on('finish', () => {
-        file.close(() => resolve());  // close() is async, call cb after close completes.
+  const rpOpts = {
+    uri: 'https://api.github.com/repos/webcatalog/electron-widevinecdm/releases/latest',
+    headers: {
+      'User-Agent': 'Request-Promise',
+      Accept: 'application/vnd.github.v3+json',
+    },
+    json: true,
+  };
+
+  return rp(rpOpts)
+    .then(({ assets }) => {
+      let matchedAsset;
+      for (let i = 0; i < assets.length; i += 1) {
+        if (assets[i].name === fileName) {
+          matchedAsset = assets[i];
+          break;
+        }
+      }
+
+      if (!matchedAsset) return Promise.reject(new Error('Cannot find valid download URL'));
+
+      return new Promise((resolve, reject) => {
+        const file = fs.createWriteStream(localZipPath);
+        https.get(matchedAsset.browser_download_url, (response) => {
+          response.pipe(file);
+          file.on('finish', () => {
+            file.close(() => resolve());  // close() is async, call cb after close completes.
+          });
+        }).on('error', (err) => { // Handle errors
+          fs.unlink(localZipPath); // Delete the file async. (But we don't check the result)
+          reject(err);
+        });
       });
-    }).on('error', (err) => { // Handle errors
-      fs.unlink(localZipPath); // Delete the file async. (But we don't check the result)
-      reject(err);
-    });
-  })
+    })
   .then(() => extractZipAsync(localZipPath, dest));
 };
 
@@ -54,6 +77,8 @@ const load = (app, dest) => {
   app.commandLine.appendSwitch('widevine-cdm-path', pluginPath);
 
   app.commandLine.appendSwitch('widevine-cdm-version', WIDEVINECDM_VERSION);
+
+  return fs.existsSync(pluginPath);
 };
 
 module.exports = {
